requirement.txt
numpy==1.18.4
python-crontab==2.5.1
python-dateutil==2.8.1
selenium==3.141.0
six==1.15.0
urllib3==1.25.9
